beautify spawn.py
I had some problem reading the code and I realised it is because the variable names are kinda meaningless, so I changed them, without changing the meaning.

multiplier? - there are 2 distinct multipliers
delta what? - delta_time
amount? Thats the most meaningless name for a number?
chance? it is a `threshhold`, a random threshold to be exact
amount and self.amount in the same function also confused me for a second